finshed exhasutive search

+++ src/tsp_stuff/Node.java
@@ -1,25 +1,25 @@
 package tsp_stuff;
 
 public class Node {
-	private double x;
-	private double y;
-	private double n;
-	public double getX() {
+	private int x;
+	private int y;
+	private int n;
+	public int getX() {
 		return x;
 	}
-	public void setX(double x) {
+	public void setX(int x) {
 		this.x = x;
 	}
-	public double getY() {
+	public int getY() {
 		return y;
 	}
-	public void setY(double y) {
+	public void setY(int y) {
 		this.y = y;
 	}
-	public double getN() {
+	public int getN() {
 		return n;
 	}
-	public void setN(double n) {
+	public void setN(int n) {
 		this.n = n;
 	}
